Remove PayNow libraries and related code. Fix an issue to commit payment result.

+++ PhotoLibrary/Class/Utils/WeChatPay/WeChatPayManager.h
@@ -14,6 +14,6 @@ typedef void (^WeChatPayCompletionHandler)(PAYRESULT payResult);
 
 + (instancetype)sharedInstance;
 
-- (void)startWeChatPayWithOrderNo:(NSString *)orderNo price:(double)price completionHandler:(WeChatPayCompletionHandler)handler;
+- (void)startWeChatPayWithOrderNo:(NSString *)orderNo price:(NSUInteger)price completionHandler:(WeChatPayCompletionHandler)handler;
 - (void)sendNotificationByResult:(PAYRESULT)result;
 @end
